+++ library/Freckle/App/Bugsnag.hs
@@ -27,9 +27,7 @@ import Data.Bugsnag (App (..), Event (..), defaultApp)
 import Data.Bugsnag.Settings (Settings (..), defaultSettings)
 import Data.List (isInfixOf)
 import Freckle.App.Async (async)
-import Freckle.App.Bugsnag.CallStack (callStackBeforeNotify)
 import Freckle.App.Bugsnag.HttpException (httpExceptionBeforeNotify)
-import Freckle.App.Bugsnag.MetaData (metaDataAnnotationsBeforeNotify)
 import Freckle.App.Bugsnag.SqlError (sqlErrorBeforeNotify)
 import qualified Freckle.App.Env as Env
 import Network.Bugsnag hiding (notifyBugsnag, notifyBugsnagWith)
@@ -111,8 +109,6 @@ envParseBugsnagSettings =
 
 globalBeforeNotify :: BeforeNotify
 globalBeforeNotify =
-  callStackBeforeNotify
-    <> metaDataAnnotationsBeforeNotify
-    <> sqlErrorBeforeNotify
+  sqlErrorBeforeNotify
     <> httpExceptionBeforeNotify
     <> maskErrorHelpers